# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PharmSci E-Voting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ballot:
    get:
      operationId: getBallot
      tags: [ballot]
      summary: Get ballot with all offices and candidates
      responses:
        "200":
          description: Ballot data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ballot"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Already voted or not voting phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ballot/submit:
    post:
      operationId: submitBallot
      tags: [ballot]
      summary: Submit completed ballot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BallotSubmission"
      responses:
        "200":
          description: Ballot submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid ballot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Already voted or not voting phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Admin auth ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ApiError:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required: [error]

    # ── Election ────────────────────────────────────────────────────────────────
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
        officeId:
          type: integer
        fullName:
          type: string
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
      required: [id, officeId, fullName]
    Office:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        displayOrder:
          type: integer
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
      required: [id, title, displayOrder, candidates]
    Ballot:
      type: object
      properties:
        offices:
          type: array
          items:
            $ref: "#/components/schemas/Office"
        voterName:
          type: string
        voterLevel:
          type: string
      required: [offices, voterName, voterLevel]
    BallotVote:
      type: object
      properties:
        officeId:
          type: integer
        candidateId:
          type: integer
      required: [officeId, candidateId]
    BallotSubmission:
      type: object
      properties:
        votes:
          type: array
          items:
            $ref: "#/components/schemas/BallotVote"
      required: [votes]

    # ── Admin session ───────────────────────────────────────────────────────────
...