# lib/api-zod/src/generated/types/ballotVote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmSci E-Voting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BallotVote {
  officeId: number;
  candidateId: number;
}